Near-Optimal Solutions
Enhanced original hypatia main.py script, integrating it with machine learning, allowing it to modify the input parameters automatically based on the results of the previous run. This will lead to a decision-space of sub-optimal solutions within a pre-defined cost slack, to be explored, instead of a single optimal solution.